[fix] 細かい修正

+++ app/controllers/comments_controller.rb
@@ -1,4 +1,6 @@
 class CommentsController < ApplicationController
+	before_action :authenticate_user!
+
 	def create
 		book = Book.find(params[:book_id])
 		# comment = Comment.new(comment_params)
@@ -6,14 +8,16 @@ class CommentsController < ApplicationController
 		# comment.book_id = book.id
 		# comment.save
 		Comment.create(user_id: current_user.id, book_id:  book.id, comment: comment_params[:comment])
-		redirect_to book_path(book)
+		redirect_to request.referer
 	end
 
 	def destroy
-		book = Book.find(params[:book_id])
 		comment = Comment.find(params[:id])
+		if comment.user != current_user
+			redirect_to request.referer
+		end
 		comment.destroy
-		redirect_to book_path(book)
+		redirect_to request.referer
 	end
 
 	private
